+++ src/game/server/tf/bot/behavior/tf_bot_get_ammo.h
@@ -12,8 +12,11 @@ class CTFBotGetAmmo : public Action< CTFBot >
 {
 public:
 	CTFBotGetAmmo( void );
+	CTFBotGetAmmo( bool crumpkin );
 
 	static bool IsPossible( CTFBot *me );			// return true if this Action has what it needs to perform right now
+	static bool IsCrumpkinPossible( CTFBot *me );
+	static bool IsSpellPossible( CTFBot *me );
 
 	virtual ActionResult< CTFBot >	OnStart( CTFBot *me, Action< CTFBot > *priorAction );
 	virtual ActionResult< CTFBot >	Update( CTFBot *me, float interval );
@@ -32,6 +35,8 @@ private:
 	PathFollower m_path;
 	CHandle< CBaseEntity > m_ammo;
 	bool m_isGoalDispenser;
+	bool m_isGoalCrumpkin;
+	bool m_isGoalSpell;
 };