+++ src/game/server/tf/bot/behavior/tf_bot_tactical_monitor.h
@@ -7,6 +7,7 @@
 #define TF_BOT_TACTICAL_MONITOR_H
 
 class CObjectTeleporter;
+class CSpellPickup;
 
 class CTFBotTacticalMonitor : public Action< CTFBot >
 {
@@ -38,6 +39,9 @@ private:
 	bool ShouldOpportunisticallyTeleport( CTFBot *me ) const;
 	CObjectTeleporter *FindNearbyTeleporter( CTFBot *me );
 	CountdownTimer m_findTeleporterTimer;
+
+	bool ShouldOpportunisticallyCollectCrumpkins( CTFBot* me ) const;
+	bool ShouldOpportunisticallyCollectSpell( CTFBot* me ) const;
 
 	void AvoidBumpingEnemies( CTFBot *me );
 };