+++ CHANGELOG.md
@@ -20,7 +20,7 @@
 tf_bot_quota_use_presets, nav_generate_noreload, tf_bot_add preset <presetname>, 
 nav_generate_auto, nav_generate_auto_view_distance, tf_mvm_popfile_requested, 
 mp_restartblock, tf_roundstarttalk_disable, tf_gamemode_override, 
-cl_default_networking_off, nav_save_compressed
+cl_default_networking_off, nav_save_compressed, tf_bot_spells
 - New inputs for tf_gamerules:  
 -- SoloSaveData, SoloUnlockItem<string>, SoloUnlockItemID<int>, SoloAddCredits<int>
 - The item schema is now loaded from items_custom.txt, allowing for extension and modularity
@@ -33,6 +33,8 @@ cl_default_networking_off, nav_save_compressed
 - Added nav files for many existing maps
 - Improved nav meshes in many existing maps
 - Saved nav files now get compressed by default
+- Bots can now proactively seek out spell and crumpkin pickups
+- Bots can now use spellbook spells
 - Fixed generated nav meshes for workshop maps not being loaded
 - Fixed nav_generate nav mesh generation not taking into account multiple spawn areas and stages
 - Fixed bots not working correctly in PLR